Replace devtools bridge with new hook
