+++ devtools/src/index.js
@@ -1,0 +1,3 @@
+import { initDevTools } from './devtools';
+
+initDevTools();
